Use WEBVIEW_HEADER instead of custom webview.h
webview.h is replaced with WEBVIEW_HEADER, which provides the
same function. It also will update with newer versions of webview.

webview.cc is added instead, which simply includes webview.h to
tell g++ that it should be compiled as an object rather than a
precompiled header (same as in the webview repo).